+++ backend/src/main/java/io/mixeway/mixewayflowapi/api/teamfindings/controller/FindingsByTeamController.java
@@ -65,9 +65,9 @@ public class FindingsByTeamController {
 
     @PreAuthorize("hasAuthority('USER')")
     @PostMapping(value= "/api/v1/teamfindings/{id}/supress")
-    public ResponseEntity<StatusDTO> supressTeamFindingList(@PathVariable("id") Long id, @RequestBody List<Long> findingIds, Principal principal){
+    public ResponseEntity<StatusDTO> supressTeamFindingList(@PathVariable("id") Long id, @RequestBody List<Long> findingIds, @RequestParam(value = "days", required = false) Integer days, Principal principal){
         try {
-            return new ResponseEntity<>(findingsByTeamService.supressTeamFindingBulk(id,findingIds,principal), HttpStatus.OK);
+            return new ResponseEntity<>(findingsByTeamService.supressTeamFindingBulk(id,findingIds,days,principal), HttpStatus.OK);
         } catch (Exception e){
             return new ResponseEntity<>( HttpStatus.BAD_REQUEST);
         }
@@ -75,9 +75,9 @@ public class FindingsByTeamController {
 
     @PreAuthorize("hasAuthority('USER')")
     @GetMapping(value= "/api/v1/teamfindings/{id}/supress/{finding}/reason/{reason}")
-    public ResponseEntity<StatusDTO> supressTeamFinding(@PathVariable("id") Long id, @PathVariable("finding") Long findingId, @PathVariable("reason") String reason, Principal principal){
+    public ResponseEntity<StatusDTO> supressTeamFinding(@PathVariable("id") Long id, @PathVariable("finding") Long findingId, @PathVariable("reason") String reason, @RequestParam(value = "days", required = false) Integer days, Principal principal){
         try {
-            return new ResponseEntity<>(findingsByTeamService.supressTeamFinding(id,findingId,reason,principal), HttpStatus.OK);
+            return new ResponseEntity<>(findingsByTeamService.supressTeamFinding(id,findingId,reason,days,principal), HttpStatus.OK);
         } catch (Exception e){
             return new ResponseEntity<>( HttpStatus.BAD_REQUEST);
         }